+++ test/algorithm/Jamfile
@@ -1,18 +1,12 @@
 # Boost.GIL (Generic Image Library) - tests
 #
-# Copyright (c) 2018 Mateusz Loskot <mateusz at loskot dot net>
+# Copyright (c) 2018-2019 Mateusz Loskot <mateusz at loskot dot net>
 #
 # Distributed under the Boost Software License, Version 1.0.
 # (See accompanying file LICENSE_1_0.txt or
 # copy at http://www.boost.org/LICENSE_1_0.txt)
 
 import testing ;
-
-project
-    : requirements
-    <include>$(BOOST_ROOT)
-    <include>..
-    ;
 
 run for_each_pixel.cpp ;
 run std_fill.cpp ;
