+++ test/channel/Jamfile
@@ -10,7 +10,6 @@ import testing ;
 
 project
     : requirements
-    <include>$(BOOST_ROOT)
     <include>..
     <library>/boost/test//boost_unit_test_framework
     <link>shared:<define>BOOST_TEST_DYN_LINK=1
