+++ test/channel/algorithm_channel_convert.cpp
@@ -10,7 +10,7 @@
 #include <cstdint>
 
 #define BOOST_TEST_MODULE test_algorithm_channel_convert
-#include <gil_test_common.hpp>
+#include "unit_test.hpp"
 #include "channel_test_fixture.hpp"
 
 namespace gil = boost::gil;
