+++ test/channel/channel_test_fixture.cpp
@@ -5,10 +5,11 @@
 // See accompanying file LICENSE_1_0.txt or copy at
 // http://www.boost.org/LICENSE_1_0.txt
 //
-#define BOOST_TEST_MODULE test_channel_test_fixture
-#include <gil_test_common.hpp>
-#include "channel_test_fixture.hpp"
 #include <limits>
+
+#define BOOST_TEST_MODULE test_channel_test_fixture
+#include "unit_test.hpp"
+#include "channel_test_fixture.hpp"
 
 namespace fixture = boost::gil::test::fixture;
