+++ test/image/Jamfile
@@ -10,7 +10,6 @@ import testing ;
 
 project
     : requirements
-    <include>$(BOOST_ROOT)
     <include>..
     ;
