+++ test/image_view/Jamfile
@@ -10,7 +10,6 @@ import testing ;
 
 project
     : requirements
-    <include>$(BOOST_ROOT)
     <include>..
     ;
