+++ test/legacy/README.md
@@ -1,0 +1,16 @@
+# Boost.GIL Legacy Tests
+
+These are GIL's original, comprehensive, all-in-one test suits.
+
+Rules of maintenance:
+
+* Run the legacy tests as part of CI builds and regression tests.
+* Keep as reference.
+* Do NOT extend.
+* Do NOT refactor.
+* Modify ONLY if absolutely necessary (a bug found in the tests).
+
+Add new test suites, with new test cases, even if their
+functional coverage is the same as of the legacy tests.
+
+See [CONTRIBUTING.md](../../CONTRIBUTING.md).
