+++ test/legacy/error_if.cpp
