+++ test/legacy/image.cpp
