+++ test/legacy/pixel_iterator.cpp
