+++ test/legacy/recreate_image.cpp
