+++ test/legacy/sample_image.cpp
