+++ test/promote_integral.cpp
@@ -18,8 +18,6 @@
 //
 #include <boost/gil/promote_integral.hpp>
 
-#include <boost/config.hpp>
-
 #include <climits>
 #include <cstddef>
 #include <algorithm>
@@ -31,7 +29,7 @@
 #ifndef BOOST_TEST_MODULE
 #define BOOST_TEST_MODULE test_promote_integral
 #endif
-#include <gil_test_common.hpp>
+#include "unit_test.hpp"
 
 // Uncomment to enable debugging output
 //#define BOOST_GIL_TEST_DEBUG 1
